# api/openapi.yaml
openapi: 3.1.0
info:
  title: PeerGit Platform API
  version: 0.1.0
  description: Health endpoints, identity and campus administration, media, and Phase 3 project/team/recruitment workflows.
servers:
  - url: https://peergit.localhost
paths:
...
  /api/v1/media:
    post:
      operationId: uploadProfileImage
      summary: Upload a bounded PNG/JPEG image that is decoded and re-encoded before storage
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                visibility: { type: string, enum: [private, campus] }
      responses:
        '201': { description: Verified media metadata }
        '413': { description: Upload exceeds 5 MiB }
        '415': { description: Invalid or unsupported image }
        '429': { description: Local upload capacity is full }
  /api/v1/media/{id}:
    get:
      operationId: downloadProfileImage
      summary: Download owner media or consented same-campus media after current authorization checks
      security: [{ sessionCookie: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Verified raster image bytes }
        '404': { description: Media is absent or not visible to this account }
...
components:
  parameters:
    CsrfToken:
      in: header
      name: X-CSRF-Token
      required: true
      schema: { type: string }
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-peergit_session
...